requirements.txt.txt
pynput  
Pillow  
requests  
pyinstaller  
